Changed dataclasses, home now uses them as well
