+++ frontend/src/classes/Group.ts
@@ -1,0 +1,7 @@
+export default class Group {
+  constructor(
+    public name: string,
+    public members: string[],
+    public id?: string,
+  ) {}
+}
