+++ frontend/src/classes/Score.ts
@@ -1,0 +1,22 @@
+import Problem from '@/classes/Problem';
+
+export default class Score {
+  constructor(
+    public problem: Problem,
+    public currentPoints: number = 0,
+    public hintUsed: boolean = false,
+    public submittedAnswer?: string,
+    public submittedCode?: string,
+    public scoreId?: string,
+  ) {}
+
+  get state() {
+    if (this.currentPoints === 0) {
+      return 'Not Done';
+    } else if (this.currentPoints === this.problem.points) {
+      return 'Strike (Max Score)';
+    } else {
+      return 'Split (Partial Score)';
+    }
+  }
+}
